backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.0.0
django-filter>=23.0
openpyxl>=3.1.0
reportlab>=4.0.0
